backend/requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
transformers==4.29.2
torch==2.2.0
python-dotenv==0.21.0
openai==1.42.0
Flask_Cors==4.0.1
